WilsonMG: Move multigrid class to separate file
